Further improvements and bug-fixes to the C code codecs
This includes support for trigger notifications (trigger[2]), and an
overhaul of the implicit return mode (formerly "call-stack"), including
the addition of the irfail and irdepth fields, and additional comments
being more consistent with the naming and nomenclature.

These files are intended as reference example implementations in C for
both the trace-encoder and the trace-decoder, providing the "heart"
of the respective algorithms.

Signed-off-by: Sean McGoogan <Sean.McGoogan@ultrasoc.com>