Further improvements and bug-fixes to the C code codecs
The main change is to the error handling and recovery, specifically
callers of unrecoverable_error() should now assume it *does* return,
and it may return immediately to its caller, et. seq. In addition, this
commit includes support for the "encoder-mode" being sent as a first-class
field in a sync+support packet. Finally, there are a few janitorial
changes to the data structures (including fewer anonymous structures).

These files are intended as reference example implementations in C for
both the trace-encoder and the trace-decoder, providing the "heart"
of the respective algorithms.

Signed-off-by: Sean McGoogan <Sean.McGoogan@ultrasoc.com>