Further improvements and bug-fixes to the C code codecs
The most visible (incompatible) impact with this commit is a change to the
API function calls te_open_trace_*(), which now include a few additional
function pointers. As a result, the previously externally-visible
user-defined functions may now be replaced by static functions, and may
also be different for each instance of the encoder/decoder state.

In addition, there is now initial support for "custom-instructions".
Including some example code on how this may be used to map custom
instructions to disassembled output generated from a custom disassembler,
for example, using "objdump -d".

There are some bug fixes pertaining to dealing with exceptions, and
the ability to generate a useful diagnostic when an exception packet
is received.  Finally, there are also a few minor additions to the data
structures, and a general improvement with some of the comments.

These files are intended as reference example implementations in C for
both the trace-encoder and the trace-decoder, providing the "heart"
of the respective algorithms.

Signed-off-by: Sean McGoogan <Sean.McGoogan@ultrasoc.com>