Further improvements and fixes to the C code codecs
This commit is mainly some minor bug-fixes, and a few miscellaneous
minor changes. These changes are:

    * Initialize decoder->decoded_cache[] properly when PC == 0x0

    * Remove the "with_address" field from the te_inst_t structure.

    * Added a new predicate function te_is_with_address()

    * Added the prefix "i" to some identifiers, making it clear this
      identifier is specific to instruction trace.

    * Updated copyright information, assigning copyright to Siemens.

    * Added SPDX copyright and licence designators.

These files are intended as reference example implementations in C for
both the trace-encoder and the trace-decoder, providing the "heart"
of the respective algorithms.

Note: any existing code will need to modified when re-basing to use this
commit. To be clear, any extant code that initializes the "with_address"
field in the struct te_inst_t can simply be deleted. There is no need
do anything else, or write any new code.

Signed-off-by: Sean McGoogan <Sean_McGoogan@mentor.com>